requirements.txt
pandas
tqdm
Pillow
requests
together
python-dotenv 